move active-counter in separate file
