+++ src/processing/subtree_queue.cpp
@@ -23,7 +23,7 @@
 #include "subtree_queue.h"
 
 #include <items/item_methods.h>
-
+#include <processing/active_counter.h>
 #include <libKitsunemimiPersistence/logger/logger.h>
 
 namespace Kitsunemimi
